minor changes, some comments

+++ src/types.rs
@@ -41,7 +41,7 @@ pub struct AppConfig {
     pub postfix_bytes: i32,
     pub keep_postfix: bool,
 
-    pub max_length_bytes: u16,
+    pub max_length_bytes: i32,
 
     pub timestamp_setting: TimestampSetting,
     pub timestamp_def: TimestampDef,
